[UI] Add ability to store user sessions
The are placed in StoredSessions folder 6-hex file names, links
are looking like this:
http://localhost:3000/1561c1

to achive this a major refactoring of the frontend was done, splitting
main page.js into session context and explorer content, also adding
session id instances of a unique page and route.
